+++ mission-control/app.ts
@@ -30,6 +30,7 @@ function makeid(length) {
 
  function blankMnmsData(d) {
      let out = JSON.parse(JSON.stringify(d))
+     out.External = []
      out.Switches.forEach((s) => {
         s.Child = null
         s.Timer = null
@@ -92,28 +93,59 @@ export = function(LocalOptions) {
         ws._data = {
             auth : false
         }
+        ws.on("close", () => {
+            console.log("Connection close: ",ws._data)
+            let sw = MnmsData[ws._data.Info.ServiceClass].findIndex(k => k.UID == ws._data.UID)
+            if(sw != -1) {
+                console.log("Found at " + sw + " deleting")
+                MnmsData[ws._data.Info.ServiceClass].splice(sw,1)
+            }       
+            ws._data = {
+                auth: false
+            }
+            
+        })
         ws.on('message', function incoming(message) {
             let node = JSON.parse(message)
             if(node.Type == "auth") {
                 if(node.Challenge == MnmsData.Challenge) {
                     ws._data.auth = true
                     console.log("new client Auth")
+                    if(!MnmsData[node.Info.ServiceClass]) MnmsData[node.Info.ServiceClass] = []
+                    let sw = MnmsData[node.Info.ServiceClass].filter(k => k.UID == node.Info.id)
+                    if(sw.length == 1) {
+                        let t = new Date
+                        sw[0].Timer = t.getTime()
+                        //console.log(node.id,MnmsData.Switches.filter(k => k.UID == node.id)[0].Timer)
+                    }
+                    else {
+                        console.log("Could not find id =",node.Info.id)
+                        let sw = MnmsData[node.Info.ServiceClass].push({
+                            IP: node.IP,
+                            Type: node.Info.Type,
+                            Ws: ws,
+                            UID: node.Info.id,
+                            Info: node.Info.Info
+                        })
+                    }
+
+
+                    ws._data.UID = node.Info.id
+                    ws._data.Info = node.Info
                 }
                 else {
                     console.log(node.Challemge,MnmsData.Challenge)
                 }
             } 
             else if(ws._data.auth) {
+                console.log("Got a message")
                 let i = Nodes.findIndex(k => k.IP == node.IP);
-                let sw = MnmsData.Switches.filter(k => k.UID == node.id)
-                if(sw.length == 1) {
-                    let t = new Date
-                    sw[0].Timer = t.getTime()
-                    //console.log(node.id,MnmsData.Switches.filter(k => k.UID == node.id)[0].Timer)
-                }
-                else {
-                    console.log("Could not find id =",node.id)
-                }
+                let sw = MnmsData[ws._data.Info.ServiceClass].filter(k => k.UID == ws._data.Info.id)
+                    if(sw.length == 1) {
+                        let t = new Date
+                        sw[0].Timer = t.getTime()
+                        //console.log(node.id,MnmsData.Switches.filter(k => k.UID == node.id)[0].Timer)
+                    }
                 if(i == -1) {
                     Nodes.push(
                         {
@@ -131,6 +163,7 @@ export = function(LocalOptions) {
                     )
                     i = Nodes.findIndex(k => k.IP == node.IP);
                 }
+                console.log("Merge now...")
                 mergeNodes(i,node,"")
                 calculateInterConnect()
             }
@@ -430,21 +463,24 @@ export = function(LocalOptions) {
     //------------------
     var MnmsData = {
         Type: "MnmsData",
+        Schema: 1,
         Workspace: "Mnms - Network Name",
         CurrentTime: 0,
         Challenge: makeid(20),
         OkSwitches: 0,
         Switches : [],
+        External: [],
         Mdns: mdns_data
     }
 
     var Datastore = require('nedb')
     , db = new Datastore({ filename: path.join(__dirname, Options.database), autoload: true });
-    db.find({ Type: "MnmsData"}, (err, docs) => {
+    db.find({ Type: "MnmsData", Schema: 1}, (err, docs) => {
         console.log(docs)
         if(docs.length==1) {
             MnmsData = docs[0]
             MnmsData.Mdns = mdns_data
+            if(!MnmsData.External) MnmsData.External = []
         }
     })
     
@@ -460,15 +496,17 @@ export = function(LocalOptions) {
         else {
             let type = ServiceOptions.Name.split(":")[0]
             let action = ServiceOptions.Name.split(":")[1]
-            if(action == "start") {
-                child_info = spawn("node",[ServicesDirectory[type],"-i",ServiceOptions.Params.IP,"-k",MnmsData.Challenge,"-y",ServiceOptions.UID ])
-                child_info.on("error",() => {
-                    child_info.kill()
-                })
-            }
-            else if(action == "stop") {
-                if(ServiceOptions.Params.Child.kill) ServiceOptions.Params.Child.kill()
-                child_info = null;
+            if(type == "cisco_switch") {
+                if(action == "start") {
+                    child_info = spawn("node",[ServicesDirectory[type],"-i",ServiceOptions.Params.IP,"-k",MnmsData.Challenge,"-y",ServiceOptions.UID ])
+                    child_info.on("error",() => {
+                        child_info.kill()
+                    })
+                }
+                else if(action == "stop") {
+                    if(ServiceOptions.Params.Child.kill) ServiceOptions.Params.Child.kill()
+                    child_info = null;
+                }
             }
         }
         return child_info
